Add files via upload
We utilized the "COVID-19 Health Indicators, Mortality (by week)" dataset to conduct an analysis on mortality data for the year 2023. Our focus was to extract figures pertaining to all causes of death over a span of 20 weeks across ten selected countries. This dataset offers a comprehensive view of weekly mortality statistics, which is crucial for assessing the overall health impacts during the specified period. The primary aim of our query was to gather and compare the total number of deaths from all causes in each country, providing insights into the health trends and potential anomalies in mortality rates. The findings from this analysis are intended to contribute to the understanding of global health trends and assist in the formulation of targeted public health responses and resource distribution.